Avatar in Glass Friend window
Adds your current avatar to the top of the Glass friend window.
It creates the object with GlassLive::createBlockhead()
It updates your avatar when GlassLive::updateBlockhead() is called.
Currently the way that it initially sets up is super weird and for some
reason needs a long delay before it actually is ready. It's currently
checking every time the overlay is opened for if it's okay to load.
Needs to be fixed.