Create an update checker for Tempered Glass
Create an update checker that runs on launch of the game. Ping lakeys.github.io/tempered_glass/repository.txt, parsing the JSON data to check for an update to Tempered Glass, should an update need to be pushed.

Include an option for the user to hide notifications for the corresponding update, saving this configuration in Glass settings.

This is a method to push updates to Tempered Glass, if necessary, without replacing the official repository directly.